refactor: renaming NowPlayingItem to QueueItem and put it in separate namespace
