+++ Stopify.Presentation/Views/NowPlayingView/NowPlayingView.xaml
@@ -2,6 +2,7 @@
     x:Class="Stopify.Presentation.Views.NowPlayingView.NowPlayingView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:Queue="clr-namespace:Stopify.Presentation.Views.QueueView"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:local="clr-namespace:Stopify.Presentation.Views.NowPlayingView"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
@@ -450,7 +451,7 @@
                         </DockPanel>
 
                         <!--  Song  -->
-                        <local:NowPlayingItem x:Name="NextSongItem" DataContext="{Binding NextSong}" />
+                        <Queue:QueueItem x:Name="NextSongItem" DataContext="{Binding NextSong}" />
                     </StackPanel>
                 </Border>
             </StackPanel>
