+++ Stopify.Presentation/Views/QueueView/QueueItem.xaml
@@ -1,9 +1,9 @@
 ﻿<UserControl
-    x:Class="Stopify.Presentation.Views.NowPlayingView.NowPlayingItem"
+    x:Class="Stopify.Presentation.Views.QueueView.QueueItem"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="clr-namespace:Stopify.Presentation.Views.NowPlayingView"
+    xmlns:local="clr-namespace:Stopify.Presentation.Views.QueueView"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d">
     <Button
