+++ Stopify.Presentation/Views/QueueView/QueueView.xaml
@@ -1,10 +1,10 @@
 ﻿<UserControl
-    x:Class="Stopify.Presentation.Views.QueueView"
+    x:Class="Stopify.Presentation.Views.QueueView.QueueView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:NowPlaying="clr-namespace:Stopify.Presentation.Views.NowPlayingView"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="clr-namespace:Stopify.Presentation.Views"
+    xmlns:local="clr-namespace:Stopify.Presentation.Views.QueueView"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d">
     <Grid VerticalAlignment="Top">
@@ -100,7 +100,7 @@
                         FontWeight="Bold"
                         Foreground="WhiteSmoke"
                         Text="Now playing" />
-                    <NowPlaying:NowPlayingItem />
+                    <local:QueueItem />
                 </StackPanel>
 
                 <!--  Next from  -->
@@ -132,16 +132,16 @@
                                 TextWrapping="Wrap" />
                         </Button>
                     </StackPanel>
-                    <NowPlaying:NowPlayingItem />
-                    <NowPlaying:NowPlayingItem />
-                    <NowPlaying:NowPlayingItem />
-                    <NowPlaying:NowPlayingItem />
-                    <NowPlaying:NowPlayingItem />
-                    <NowPlaying:NowPlayingItem />
-                    <NowPlaying:NowPlayingItem />
-                    <NowPlaying:NowPlayingItem />
-                    <NowPlaying:NowPlayingItem />
-                    <NowPlaying:NowPlayingItem />
+                    <local:QueueItem />
+                    <local:QueueItem />
+                    <local:QueueItem />
+                    <local:QueueItem />
+                    <local:QueueItem />
+                    <local:QueueItem />
+                    <local:QueueItem />
+                    <local:QueueItem />
+                    <local:QueueItem />
+                    <local:QueueItem />
                 </StackPanel>
             </StackPanel>
         </ScrollViewer>
